extract createConfig() to another file to be reused
