+++ rollup.create-config.js
@@ -1,0 +1,41 @@
+import buble from 'rollup-plugin-buble';
+import json from 'rollup-plugin-json';
+import resolve from 'rollup-plugin-node-resolve';
+import commonjs from 'rollup-plugin-commonjs';
+
+const ensureArray = maybeArr => Array.isArray(maybeArr) ? maybeArr : [maybeArr];
+
+const createConfig = (opts) => {
+	opts = opts || {};
+	const browser = opts.browser || false;
+	const external = opts.external || ['acorn', 'magic-string'];
+	const output = ensureArray(opts.output);
+
+	return {
+		input: 'src/index.js',
+		output: output.map(format => Object.assign({}, format, {
+			name: 'buble',
+			sourcemap: true
+		})),
+		external: external,
+		plugins: [
+			json(),
+			commonjs(),
+			buble({
+				target: !browser ? { node: 4 } : null,
+				include: [
+					'src/**',
+					'node_modules/regexpu-core/**',
+					'node_modules/unicode-match-property-ecmascript/**',
+					'node_modules/unicode-match-property-value-ecmascript/**',
+				],
+				transforms: {
+					dangerousForOf: true
+				}
+			}),
+			resolve()
+		],
+	};
+};
+
+export default createConfig;
